tried a different message in recv, MSG_BATCH, seems to work now!

+++ oppgave_5/client/include/http_utils.h
@@ -8,7 +8,7 @@
 
 struct HTTP_RESPONSE {
     int iStatusCode;
-    int iContentLength;
+    long iContentLength;
     char szStatusMessage[56];
     char szVersion[56];
     char szContentType[56];
